Reorganized benchmarker in files
