+++ hgp_lib/benchmarkers/gp_benchmarker.py
@@ -1,12 +1,13 @@
 import os
-import random
-from typing import Callable, List, NamedTuple
+from typing import Callable, List
 
-import numpy as np
 from numpy import ndarray
-from sklearn.model_selection import StratifiedKFold, train_test_split
 from tqdm import tqdm
 from tqdm.contrib.concurrent import process_map
+
+from .config import BenchmarkConfig
+from .results import aggregate_results
+from .runner import execute_single_run, single_run_wrapper
 
 from ..crossover import CrossoverExecutor
 from ..metrics import BenchmarkMetrics, RunMetrics
@@ -14,179 +15,11 @@ from ..mutations import MutationExecutor
 from ..populations import PopulationGenerator
 from ..rules import Rule
 from ..selections import BaseSelection
-from ..trainers import GPTrainer
-from ..utils.metrics import optimize_scorer_for_data
 from ..utils.validation import (
     check_isinstance,
     validate_callable,
     validate_trainer_params,
 )
-
-
-class BenchmarkConfig(NamedTuple):
-    data: ndarray
-    labels: ndarray
-    test_size: float
-    n_folds: int
-    num_epochs: int
-    score_fn: Callable[[ndarray, ndarray], float]
-    val_score_fn: Callable[[ndarray, ndarray], float]
-    optimize_scorer: bool
-    check_valid: Callable[[Rule], bool] | None
-    population_generator: PopulationGenerator | None
-    mutation_executor: MutationExecutor | None
-    crossover_executor: CrossoverExecutor | None
-    selection: BaseSelection | None
-    regeneration: bool
-    regeneration_patience: int
-    val_every: int
-    show_fold_progress: bool
-    show_epoch_progress: bool
-    progress_bar: bool
-
-
-def _execute_single_run(
-    run_id: int,
-    seed: int,
-    config: BenchmarkConfig,
-) -> RunMetrics:
-    """
-    Execute one benchmark run: stratified train/test split, k-fold CV, select best fold, test.
-
-    This is a module-level function so it can be pickled for process_map.
-
-    Args:
-        run_id (int):
-            Index of the run (0-based).
-        seed (int):
-            Random seed for stratified split and k-fold.
-        config (BenchmarkConfig):
-            Configuration for the benchmark run.
-
-    Returns:
-        RunMetrics: Metrics for the run including fold scores, best rule, and validation score.
-
-    Raises:
-        RuntimeError: If no best rule is available after training a fold.
-    """
-    # Compute effective progress bar flags
-    show_folds = config.show_fold_progress and config.progress_bar
-    show_epochs = config.show_epoch_progress and config.progress_bar
-
-    np.random.seed(seed)
-    random.seed(seed)
-
-    train_data, test_data, train_labels, test_labels = train_test_split(
-        config.data,
-        config.labels,
-        test_size=config.test_size,
-        stratify=config.labels,
-        random_state=seed,
-    )
-
-    skf = StratifiedKFold(n_splits=config.n_folds, shuffle=True, random_state=seed)
-
-    fold_train_scores: List[float] = []
-    fold_val_scores: List[float] = []
-    fold_best_rules: List[Rule] = []
-
-    fold_iterator = list(skf.split(train_data, train_labels))
-    if show_folds:
-        fold_iterator = tqdm(
-            fold_iterator, total=config.n_folds, desc="  Folds", leave=False
-        )
-
-    for train_idx, val_idx in fold_iterator:
-        fold_train = train_data[train_idx]
-        fold_val = train_data[val_idx]
-        fold_train_labels = train_labels[train_idx]
-        fold_val_labels = train_labels[val_idx]
-
-        trainer = GPTrainer(
-            score_fn=config.score_fn,
-            num_epochs=config.num_epochs,
-            train_data=fold_train,
-            train_labels=fold_train_labels,
-            val_data=fold_val,
-            val_labels=fold_val_labels,
-            val_score_fn=config.val_score_fn,
-            check_valid=config.check_valid,
-            population_generator=config.population_generator,
-            mutation_executor=config.mutation_executor,
-            crossover_executor=config.crossover_executor,
-            selection=config.selection,
-            regeneration=config.regeneration,
-            regeneration_patience=config.regeneration_patience,
-            val_every=config.val_every,
-            progress_bar=show_epochs,
-            progress_desc="    Epochs" if show_epochs else None,
-            optimize_scorer=config.optimize_scorer,
-        )
-        trainer.fit()
-
-        train_metrics = trainer.gp_algo.validate_best(
-            trainer.train_data,
-            trainer.train_labels,
-            score_fn=trainer.score_fn,
-            all_time_best=True,
-        )
-        val_metrics = trainer.gp_algo.validate_best(
-            trainer.val_data,
-            trainer.val_labels,
-            score_fn=trainer.val_score_fn,
-            all_time_best=True,
-        )
-
-        fold_train_scores.append(float(train_metrics["best"]))
-        fold_val_scores.append(float(val_metrics["best"]))
-
-        if trainer.gp_algo.real_best_rule is None:
-            raise RuntimeError(
-                f"No best rule available after training fold. "
-                f"Run {run_id}, seed {seed}."
-            )
-        fold_best_rules.append(trainer.gp_algo.real_best_rule.copy())
-
-    best_fold_idx = int(np.argmax(fold_val_scores))
-    best_fold_val_score = fold_val_scores[best_fold_idx]
-    best_rule = fold_best_rules[best_fold_idx]
-
-    if config.optimize_scorer:
-        test_score_fn, test_data_opt, test_labels_opt = optimize_scorer_for_data(
-            config.score_fn, test_data, test_labels
-        )
-    else:
-        test_score_fn = config.score_fn
-        test_data_opt = test_data
-        test_labels_opt = test_labels
-
-    test_predictions = best_rule.evaluate(test_data_opt)
-    test_score = float(test_score_fn(test_predictions, test_labels_opt))
-
-    return RunMetrics(
-        run_id=run_id,
-        seed=seed,
-        fold_train_scores=fold_train_scores,
-        fold_val_scores=fold_val_scores,
-        best_fold_idx=best_fold_idx,
-        best_fold_val_score=best_fold_val_score,
-        test_score=test_score,
-        best_rule=best_rule,
-    )
-
-
-def _single_run_wrapper(args: tuple) -> RunMetrics:
-    """
-    Picklable wrapper for process_map.
-
-    Args:
-        args (tuple): Tuple of (run_id, seed, config).
-
-    Returns:
-        RunMetrics: Metrics for the run.
-    """
-    run_id, seed, config = args
-    return _execute_single_run(run_id, seed, config)
 
 
 class GPBenchmarker:
@@ -494,62 +327,10 @@ class GPBenchmarker:
             disable=not (self.progress_bar and self.show_run_progress),
         ):
             seed = self.base_seed + run_id
-            metrics = _execute_single_run(run_id, seed, config)
+            metrics = execute_single_run(run_id, seed, config)
             run_metrics.append(metrics)
 
         return run_metrics
-
-    def _aggregate_results(self, run_metrics: List[RunMetrics]) -> BenchmarkMetrics:
-        """
-        Aggregate per-run metrics into BenchmarkMetrics.
-
-        Args:
-            run_metrics (List[RunMetrics]): Metrics from each run.
-
-        Returns:
-            BenchmarkMetrics: Aggregated metrics containing mean/std of test and
-            validation scores, plus the full list of per-run metrics.
-
-        Examples:
-            >>> import numpy as np
-            >>> from hgp_lib.benchmarkers import GPBenchmarker
-            >>> data = np.array([[True, False], [False, True], [True, True], [False, False]])
-            >>> labels = np.array([1, 0, 1, 0])
-            >>> def score(p, l): return float((p == l).mean())
-            >>> b = GPBenchmarker(score, 1, data, labels, optimize_scorer=False)
-            >>> mock_metrics = [
-            ...     {"test_score": 0.8, "best_fold_val_score": 0.75, "run_id": 0,
-            ...      "seed": 0, "fold_train_scores": [0.7], "fold_val_scores": [0.75],
-            ...      "best_fold_idx": 0, "best_rule": None},
-            ...     {"test_score": 0.9, "best_fold_val_score": 0.85, "run_id": 1,
-            ...      "seed": 1, "fold_train_scores": [0.8], "fold_val_scores": [0.85],
-            ...      "best_fold_idx": 0, "best_rule": None},
-            ... ]
-            >>> result = b._aggregate_results(mock_metrics)
-            >>> round(result["mean_test_score"], 2)
-            0.85
-            >>> round(result["std_test_score"], 2)
-            0.05
-            >>> round(result["mean_best_val_score"], 2)
-            0.8
-            >>> len(result["all_test_scores"])
-            2
-            >>> len(result["all_best_rules"])
-            2
-        """
-        test_scores = [m["test_score"] for m in run_metrics]
-        best_val_scores = [m["best_fold_val_score"] for m in run_metrics]
-        all_best_rules = [m["best_rule"] for m in run_metrics]
-
-        return BenchmarkMetrics(
-            run_metrics=run_metrics,
-            mean_test_score=float(np.mean(test_scores)),
-            std_test_score=float(np.std(test_scores)),
-            mean_best_val_score=float(np.mean(best_val_scores)),
-            std_best_val_score=float(np.std(best_val_scores)),
-            all_test_scores=test_scores,
-            all_best_rules=all_best_rules,
-        )
 
     def fit(self) -> BenchmarkMetrics:
         """
@@ -585,7 +366,7 @@ class GPBenchmarker:
                 for run_id in range(self.num_runs)
             ]
             run_metrics = process_map(
-                _single_run_wrapper,
+                single_run_wrapper,
                 run_args,
                 max_workers=effective_n_jobs,
                 desc="Benchmark Runs",
@@ -593,4 +374,4 @@ class GPBenchmarker:
             )
 
         self._run_metrics = run_metrics
-        return self._aggregate_results(run_metrics)
+        return aggregate_results(run_metrics)
